pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ohlcv-hub"
version = "0.1.0"
description = "OHLCV data fetching and processing tool"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "OHLCV Hub Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "typer>=0.12.0",
    "python-dateutil>=2.8.0",
    "pandas_market_calendars>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "bandit>=1.7.0",
    "pip-audit>=2.7.0",
]

[project.scripts]
ohlcv-hub = "ohlcv_hub.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["ohlcv_hub*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
